requirements.txt
# ============================================================
# FIN-Advisor RAG — Dependencias de Python
# ============================================================
# Instalar con:  pip install -r requirements.txt
# ============================================================

# --- Servidor web y API ---
fastapi
uvicorn
pydantic

# --- Framework del agente ---
langchain
langchain-community

# --- Proveedores de LLM ---
langchain-google-genai
langchain-huggingface
langchain-openai
langchain-groq

# --- Base de datos vectorial ---
chromadb

# --- Modelo de embeddings ---
sentence-transformers

# --- Pruebas ---
hypothesis
pytest
pytest-asyncio
httpx

# --- Configuración ---
python-dotenv
